- Oops, the overflow mentioned by foobar was actually because the equippable base-jobs field was unsigned short when we now need it to be unsigned int. Fixed now.
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@5884 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/map/itemdb.c
@@ -194,7 +194,7 @@ struct item_data* itemdb_exists(int nameid)
  * to the format used by the map server. [Skotlex]
  *------------------------------------------
  */
-static void itemdb_jobid2mapid(unsigned short *bclass, int jobmask)
+static void itemdb_jobid2mapid(unsigned int *bclass, int jobmask)
 {
 	int i;
 	bclass[0]= bclass[1]= bclass[2]= 0;
